feat(irgen): 解决作用域变量泄露问题，修改alloca翻译时机
优化二元运算指令dump；
部分效果优化;
解决作用域变量泄露问题；
修改alloca翻译时机。

+++ src/irgen/type.cpp
@@ -18,7 +18,7 @@ std::string PrimitiveType::dump() const
     case INT:
         return "i32";
     case REAL:
-        return "double";
+        return "float";
     case BOOL:
         return "i1";
     case CHAR:
